# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SiliconPath API",
    "version": "1.0.0",
    "description": "SiliconPath - VLSI/Embedded Career Platform API"
  },
  "servers": [
    {
      "url": "https://siliconpath.vercel.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "paths": {
    "/api/opportunities": {
      "get": {
        "summary": "List opportunities",
        "description": "Paginated list of opportunities with filtering, search, and sorting",
        "operationId": "listOpportunities",
        "tags": [
          "opportunities"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eligibility",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deadline",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "All",
                "This Week",
                "This Month",
                "Later"
              ]
            }
          },
          {
            "name": "verified",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "true"
              ],
              "default": "true"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated opportunity list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create opportunity",
        "description": "Create a new opportunity (admin only)",
        "operationId": "createOpportunity",
        "tags": [
          "opportunities"
        ],
        "security": [
          {
            "AdminAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpportunityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/scrape/status": {
      "get": {
        "summary": "Scrape job status",
        "operationId": "getScrapeStatus",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "AdminAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Scrape source statuses"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Opportunity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "salary_range": {
            "type": "string"
          },
          "deadline": {
            "type": "string",
            "format": "date"
          },
          "eligibility": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "apply_url": {
            "type": "string",
            "format": "uri"
          },
          "source_url": {
            "type": "string",
            "format": "uri"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "rejected"
            ]
          },
          "is_active": {
            "type": "boolean"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "scraped",
              "employer_posted",
              "admin"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          }
        }
      },
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "government",
              "academic",
              "private"
            ]
          },
          "logo_url": {
            "type": "string",
            "format": "uri"
          },
          "website": {
            "type": "string",
            "format": "uri"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OpportunityListResponse": {
        "type": "object",
        "properties": {
          "opportunities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Opportunity"
            }
          },
          "count": {
            "type": "integer"
          },
          "total_count": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          }
        }
      },
      "OpportunityResponse": {
        "type": "object",
        "properties": {
          "opportunity": {
            "$ref": "#/components/schemas/Opportunity"
          }
        }
      },
      "CreateOpportunityRequest": {
        "type": "object",
        "required": [
          "title",
          "category"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "salary_range": {
            "type": "string"
          },
          "deadline": {
            "type": "string",
            "format": "date"
          },
          "eligibility": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "apply_url": {
            "type": "string",
            "format": "uri"
          },
          "source_url": {
            "type": "string",
            "format": "uri"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "scraped",
              "employer_posted",
              "admin"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Custom"
      },
      "CronAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}